New rule for including .c and .h files.

+++ egroot/src/pkg/runtime/noos/tasks_c.h
@@ -1,4 +1,4 @@
-extern uint32 MaxTasks;
+extern byte MaxTasks;
 
 __attribute__ ((always_inline))
 extern inline int runtime_noos_MaxTasks() {
